change point type to float

+++ src/utilities/types.py
@@ -1,3 +1,3 @@
-Point = tuple[int, int]
+Point = tuple[float, float]
 Path = list[Point]
 NetAddr = int
